+++ helpers/resize.py
@@ -4,17 +4,38 @@ from PIL import Image
 import json
 import os
 
+
+def resize_img(from_path: str, size: tuple, to_path: str = None):
+    """
+
+    Args:
+        from_path: path of image being resized
+        to_path: path of the new resized images (optional), if None, image saves in from_path
+        size: new size of image
+
+    Returns:
+        None
+
+    """
+    im = Image.open(from_path)
+    im = im.resize(size, Image.ANTIALIAS)
+    if to_path is None:
+        im.save(from_path, 'JPEG', quality=90) # rewrites the same file
+    else:
+        im.save(to_path, 'JPEG', quality=90)
+
+
 def resize_dir(load_folder: str, save_folder: str, size: tuple):
     """
         resizes images in a folder
     """
     image_paths = glob(f'{load_folder}/*jpg')
     
-    for image in tqdm(image_paths):
-        filename = os.path.basename(image)
-        im = Image.open(image)
-        im = im.resize(size, Image.ANTIALIAS)
-        im.save(f'{save_folder}/{filename}', 'JPEG', quality=90)
+    for image_path in tqdm(image_paths):
+        filename = os.path.basename(image_path)
+        save_path = f'{save_folder}/{filename}'
+        resize_img(image_path, size, save_path)
+
 
 def resize_coco_bbox_annotations(coco_json: str, from_size: tuple, size: tuple):
     """
@@ -30,3 +51,4 @@ def resize_coco_bbox_annotations(coco_json: str, from_size: tuple, size: tuple):
 
     path = os.path.splitext(coco_json)[0] + f'_{size[0]}.json' # takes path and appends _{size}.json to differenciate files
     json.dump(data,open(path, 'w'))
+