+++ helpers/testing.py
@@ -1,5 +1,7 @@
+import torchvision.transforms
 from torchvision.io import read_image
 from torchvision.utils import draw_bounding_boxes
+from torchvision.utils import make_grid
 from torchvision.ops import box_convert
 import matplotlib.pyplot as plt
 import torch
@@ -11,6 +13,32 @@ from glob import glob
 from detectors import PieceDetector, BoardDetector
 from dataloader import BoardDetectorDataset, PieceDetectorDataset
 from helpers import warp
+
+
+def piece_detector_results(idx: int,
+                                piece_detector: PieceDetector,
+                                piece_data: PieceDetectorDataset,
+                                device):
+    """
+    using draw_bounding_boxes, the function displays results of predicted boxes
+
+    @param idx: path of the image being tested
+    @param piece_data: initialized, PieceDetectorDataset
+    @param piece_detector: initialized, PieceDetector model
+    @param device: device
+
+    @return: returns two image tensors, 1st is the actual output, 2nd is the prediction
+    """
+    img, target = piece_data[idx]
+    real_boxes, real_labels, _ = target.values()
+    real_labels = [piece_data.classes[l]["name"] for l in real_labels]
+
+    boxes, labels, scores = piece_detector(img.unsqueeze(0).to(device))[0].values()
+    labels = [piece_data.classes[l]["name"] for l in labels] # convert int labels to string labels (0 -> white pawn)
+    og_img = (img * 255).to(torch.uint8)
+    out = draw_bounding_boxes(og_img, boxes, labels, width=2)
+    real_out = draw_bounding_boxes(og_img, real_boxes, real_labels, width=2)
+    return real_out, out
 
 
 def show_piece_detector_results(idx: int, 
@@ -27,32 +55,24 @@ def show_piece_detector_results(idx: int,
 
     @return: None
     """
-    img, target = piece_data[idx]
-    real_boxes, real_labels = target.values()
-    real_labels = [piece_data.classes[l]["name"] for l in real_labels]
-
-    boxes, labels, scores = piece_detector(img.unsqueeze(0).to(device))[0].values()
-    labels = [piece_data.classes[l]["name"] for l in labels] # convert int labels to string labels (0 -> white pawn)
-    og_img = (img * 255).to(torch.uint8)
-    print(len(boxes))
-    out = draw_bounding_boxes(og_img, boxes * piece_data.w, labels, width=5)
-    real_out = draw_bounding_boxes(og_img, real_boxes * piece_data.w, real_labels, width=5)
-    
-    f, ax = plt.subplots(1,2, figsize=(12, 10))
+    real_out, out = piece_detector_results(idx, piece_detector, piece_data, device)
+    f, ax = plt.subplots(1, 2, figsize=(12, 10))
     ax[0].imshow(real_out.permute(1, 2, 0))
     ax[1].imshow(out.permute(1, 2, 0))
     plt.show()
+
 
 def show_board_detector_results(idx: int, 
                                 board_detector: BoardDetector,
                                 board_data: BoardDetectorDataset,
                                 device):
+    assert board_data.target == 'points', 'BoardDetector target parameter is not set to "points"'
     """
-    Displays two board dectector results, board detector predition and actual target
+    Displays two board detector results, board detector prediction and actual target
     Args:
-        (int) idx: index of image to diplay
-        (BoardDetector object) board_detector: initalized, board detector model
-        (BoardDetectorDataset object) board_data: initalized, board dataset
+        (int) idx: index of image to display
+        (BoardDetector object) board_detector: initialized, board detector model
+        (BoardDetectorDataset object) board_data: initialized, board dataset
 
     Returns:
         None
@@ -84,4 +104,25 @@ def show_board_detector_results(idx: int,
     coords = list(zip(keypoints.tolist()[0::2], keypoints.tolist()[1::2]))
     warped = warp(img.unsqueeze(0), torch.tensor(coords).unsqueeze(0))
     ax[1, 1].imshow(warped[0].permute(1, 2, 0))
+    plt.show()
+
+
+def show_board_detector_results_masked(idx: int,
+                                       board_detector: BoardDetector,
+                                       board_data: BoardDetectorDataset,
+                                       device):
+    assert board_data.target == 'mask', 'BoardDetector target parameter is not set to "mask"'
+
+    img, mask_actual = board_data[idx]
+    mask = board_detector(img.unsqueeze(0).to(device))[0]
+
+    criteron = torch.nn.BSELoss()
+    loss = criteron(mask.to(device), mask_actual.to(device))
+    print(f'loss on displayed image: {loss}')
+
+    f, ax = plt.subplots(2,2, figsize=(12, 10))
+    ax[0, 0].imshow(img.permute(1, 2, 0))
+    ax[0, 0].imshow(mask_actual.permute(1, 2, 0), 'red', alpha=0.7)
+    ax[0, 1].imshow(img.permute(1, 2, 0))
+    ax[0, 0].imshow(mask.permute(1, 2, 0), 'red', alpha=0.7)
     plt.show()